batch: Move source snapshot management
Batch operations use per-file source commits to remember the file content
that ownership claims were created against. The source commit cache is
currently implemented as data.batch_sources.

That placement makes the batch package depend on data for its own source
snapshot machinery. The dependency reverses the package story because
session commands and batch storage both need batch-specific source
snapshots.

This commit addresses that by moving the module to batch.source_snapshots
and rewriting source imports to use the batch-owned source snapshot
boundary.

Subsequent commits will move tests and architecture guards to the new
module path.
