+++ src/git_stage_batch/batch/source_advancement.py
@@ -6,7 +6,7 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 
 from ..core.buffer import LineBuffer
-from ..data.batch_sources import create_batch_source_commit
+from .source_snapshots import create_batch_source_commit
 from ..utils.repository_buffers import (
     load_git_object_as_buffer,
     load_working_tree_file_as_buffer,
