+++ src/git_stage_batch/commands/selection/discard_line_replacement.py
@@ -34,7 +34,7 @@ from ...batch.validation import batch_exists
 from ...core.buffer import LineBuffer, buffer_ends_with_lf
 from ...core.line_selection import parse_line_selection
 from ...core.replacement import ReplacementPayload, coerce_replacement_payload
-from ...data.batch_sources import (
+from ...batch.source_snapshots import (
     create_batch_source_commit,
     load_session_batch_sources,
     save_session_batch_sources,
